requirements.txt
discord==1.7.3
discord.py[voice]==1.7.3
requests~=2.26.0
python-dotenv==0.19.2
youtube_dl==2021.12.17
ffmpeg==1.4